Projects/transport solution/backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-filter
pillow
python-dotenv
drf-yasg
django-cors-headers
